auth/requirements.txt
firebase-admin==6.2.0
streamlit-authenticator==0.2.3
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
